requirements.txt
fastapi[all]==0.99.0
pydantic==1.9
pymongo[srv]==3.11.0
python-dotenv==0.19.2
pytest==7.0.1
uvicorn==0.17.6